+++ twin4build/api/codes/ml_layer/whatif_scenario/what_if_request_api.py
@@ -1,3 +1,4 @@
+
 import os 
 import sys
 import time
@@ -7,19 +8,11 @@ import requests
 import pandas as pd
 from datetime import datetime
 
-###Only for testing before distributing package
-if __name__ == '__main__':
-    # Define a function to move up in the directory hierarchy
-    uppath = lambda _path, n: os.sep.join(_path.split(os.sep)[:-n])
-    # Calculate the file path using the uppath function
-    file_path = uppath(os.path.abspath(__file__), 5)
-    # Append the calculated file path to the system path
-    sys.path.append(file_path)
 
-else: from twin4build.utils.uppath import uppath
+from twin4build.utils.uppath import uppath
 
 # import custom modules
-from twin4build.api.codes.ml_layer.input_data import input_data
+from twin4build.api.codes.ml_layer.whatif_scenario.what_if_input_data import input_data
 from twin4build.api.codes.database.db_data_handler import db_connector
 from twin4build.config.Config import ConfigReader
 from twin4build.logger.Logging import Logging
@@ -33,16 +26,15 @@ logger = Logging.get_logger('API_logfile')
 """
 Right now we are connecting 2 times with DB that needs to be corrected.
 """
-class request_class:
+class what_if_request_class:
      
     def __init__(self):
         # Initialize the configuration, database connection, process input data, and disconnect
         logger.info("[request_to_api]: Entered initialise function")
-        self.config = self.get_configuration()
+        
+        self.config  = self.get_configuration()
         
         self.url = self.config['simulation_api_cred']['url']
-        self.history_table_to_add_data = self.config['simulation_variables']['table_to_add_data']
-        self.forecast_table_to_add_data =  self.config['forecast_simulation_variables']['table_to_add_data']
 
         self.db_handler = db_connector()
         self.db_handler.connect()
@@ -61,16 +53,17 @@ class request_class:
         try:
             self.conf = ConfigReader()
             config_path = os.path.join(os.path.abspath(
-                uppath(os.path.abspath(__file__), 4)), "config", "conf.ini")
+                uppath(os.path.abspath(__file__), 5)), "config", "conf.ini")
+            
             self.config = self.conf.read_config_section(config_path)
             logger.info("[request_class]: Configuration has been read from file")
 
             #url of web service will be placed here
             #url = self.config["simulation_api_cred"]["url"]
-            return self.config
+            return self.config 
         except Exception as e:
             logger.error("Error reading configuration: %s", str(e))
-            print(e)
+            print("Error reading configuration:",e)
 
     # this function creates json file of the object passed - used in testing
     def create_json_file(self,object,filepath):
@@ -134,8 +127,9 @@ class request_class:
         except Exception as error_creating:
             logger.error("[request_to_api] : create_dmi_forecast_key error %s",str(error_creating))
     
-    def request_to_simulator_api(self,start_time,end_time,time_with_warmup,forecast):
+    def what_if_request_to_simulator_api(self,start_time,end_time,time_with_warmup,forecast):
         try :
+
             # get data from multiple sources code wiil be called here
             logger.info("[request_class]:Getting input data from input_data class")
 
@@ -144,8 +138,7 @@ class request_class:
             # validating the inputs coming ..
             input_validater = self.validator.validate_input_data(i_data,forecast)
 
-            if forecast:
-                i_data = self.create_dmi_forecast_key(i_data)
+            i_data = self.create_dmi_forecast_key(i_data)
             
             self.create_json_file(i_data,"input_data.json")
                 
@@ -158,14 +151,10 @@ class request_class:
                 # Check if the request was successful (HTTP status code 200)
                 if response.status_code == 200:
                     model_output_data = response.json()
-                    
-                    #self.create_json_file(model_output_data,"raw_model_output.json")
 
                     response_validater = self.validator.validate_response_data(model_output_data)
                     #validating the response
                     if response_validater:
-
-                        #print("Response validator +++++++++++++++++++")
                         #filtering out the data between the start and end time ...
                         model_output_data = self.extract_actual_simulation(model_output_data,start_time,end_time)
 
@@ -176,10 +165,7 @@ class request_class:
                                     
                         self.create_json_file(input_list_data,"response_after_transformation.json")
 
-                        if not forecast: # forecast & history table names will get switched 
-                            table_to_add_data = self.history_table_to_add_data # ml_simulation_results
-                        else:
-                            table_to_add_data = self.forecast_table_to_add_data
+                        table_to_add_data = "ml_what_if_results"
 
                         self.db_handler.add_data(table_to_add_data,inputs=input_list_data)
 
@@ -205,38 +191,6 @@ class request_class:
                 logger.info("[request_to_simulator_api]:disconnect error %s",str(disconnect_error))
     
 
-if __name__ == '__main__':
-
-    request_class_obj= request_class()
-    config = request_class_obj.get_configuration()
     
-    request_timer_obj = RequestTimer(request_class_obj)
-
-    simulation_duration = int(config["simulation_variables"]["simulation_duration"])
         
-    # Schedule subsequent function calls at 1-hour intervals
-    #changing to 2 min for testing
-    #sleep_interval = 120
-    sleep_interval = simulation_duration * 60 * 60  # 1 hours in seconds
-
-    request_timer_obj.request_simulator()
-    # Create a schedule job that runs the request_simulator function every 2 hours
-    job = schedule.every(sleep_interval).seconds.do(request_timer_obj.request_simulator)
-
-    while True:
-        try :
-            schedule.run_pending()
-            print("Function called at:", time.strftime("%Y-%m-%d %H:%M:%S"))
-            logger.info("[main]:Function called at:: %s"%time.strftime("%Y-%m-%d %H:%M:%S"))
-            # Sleep for the remaining time until the next 2-hour interval
-            time.sleep(sleep_interval)
-        
-        except Exception as schedule_error:
-            schedule.cancel_job(job)
-            request_class_obj.db_handler.disconnect()
-            request_class_obj.data_obj.db_disconnect()
-            logger.error("An Error has occured: %s",str(schedule_error))
-            break
-
-        # model line 1036 , needede dmi , forecast ? 
-        # no space == history / np.isnan        
+
